+++ tests/test_operators_geometry.py
@@ -9,9 +9,7 @@ import pandas as pd
 import pytest
 from shapely.geometry import Point, Polygon
 
-from geoterminal.geometry_operations.geometry_operations import (
-    GeometryProcessor,
-)
+from geoterminal.operators.geometry_operations import GeometryProcessor
 
 
 @pytest.fixture